requirements.txt
flask==2.2.5
gunicorn==21.2.0
werkzeug==2.2.3
python-docx==0.8.11
htmldocx
PyMuPDF
reportlab
python-magic
